chore: cleanup tokens list
