+++ app/components/UI/DeFiPositions/DeFiPositionsControlBar.test.tsx
@@ -83,11 +83,7 @@ jest.mock('../../hooks/useNetworksByNamespace/useNetworksByNamespace', () => ({
   },
 }));
 
-jest.mock('../Tokens/TokensBottomSheet', () => ({
-  createTokenBottomSheetFilterNavDetails: () => [
-    'RootModalFlow',
-    { screen: 'TokenFilter' },
-  ],
+jest.mock('../Tokens/TokensBottomSheet/TokenSortBottomSheet', () => ({
   createTokensBottomSheetNavDetails: () => [
     'RootModalFlow',
     { screen: 'TokensBottomSheet' },
