+++ app/components/UI/Tokens/TokenList/TokenList.tsx
@@ -10,11 +10,10 @@ import {
 
 import { TokenI } from '../types';
 import { strings } from '../../../../../locales/i18n';
-import { TokenListItem, TokenListItemBip44 } from './TokenListItem';
+import { TokenListItem } from './TokenListItem/TokenListItem';
 import { WalletViewSelectorsIDs } from '../../../../../e2e/selectors/wallet/WalletView.selectors';
 import { useNavigation } from '@react-navigation/native';
 import Routes from '../../../../constants/navigation/Routes';
-import { selectMultichainAccountsState2Enabled } from '../../../../selectors/featureFlagController/multichainAccounts';
 import { selectHomepageRedesignV1Enabled } from '../../../../selectors/featureFlagController/homepage';
 import {
   Box,
@@ -61,14 +60,6 @@ const TokenListComponent = ({
     selectHomepageRedesignV1Enabled,
   );
 
-  // BIP44 MAINTENANCE: Once stable, only use TokenListItemBip44
-  const isMultichainAccountsState2Enabled = useSelector(
-    selectMultichainAccountsState2Enabled,
-  );
-  const TokenListItemComponent = isMultichainAccountsState2Enabled
-    ? TokenListItemBip44
-    : TokenListItem;
-
   const listRef = useRef<FlashListRef<FlashListAssetKey>>(null);
 
   const navigation = useNavigation();
@@ -101,7 +92,7 @@ const TokenListComponent = ({
 
   const renderTokenListItem = useCallback(
     ({ item }: { item: FlashListAssetKey }) => (
-      <TokenListItemComponent
+      <TokenListItem
         assetKey={item}
         showRemoveMenu={showRemoveMenu}
         setShowScamWarningModal={setShowScamWarningModal}
@@ -115,7 +106,6 @@ const TokenListComponent = ({
       setShowScamWarningModal,
       privacyMode,
       showPercentageChange,
-      TokenListItemComponent,
       isFullView,
     ],
   );
@@ -127,7 +117,7 @@ const TokenListComponent = ({
         testID={WalletViewSelectorsIDs.TOKENS_CONTAINER_LIST}
       >
         {displayTokenKeys.map((item, index) => (
-          <TokenListItemComponent
+          <TokenListItem
             key={`${item.address}-${item.chainId}-${item.isStaked ? 'staked' : 'unstaked'}-${index}`}
             assetKey={item}
             showRemoveMenu={showRemoveMenu}
