--- app/components/UI/Tokens/TokenList/TokenListItem/index.test.tsx
@@ -1,2372 +1,0 @@
-import React from 'react';
-import { render } from '@testing-library/react-native';
-import { Provider, useSelector } from 'react-redux';
-import { NavigationContainer } from '@react-navigation/native';
-import { configureStore } from '@reduxjs/toolkit';
-import { TextColor } from '../../../../../component-library/components/Texts/Text';
-import { TOKEN_RATE_UNDEFINED } from '../../constants';
-import { TokenListItem } from './index';
-import { FlashListAssetKey } from '..';
-
-// Mock dependencies
-jest.mock('@react-navigation/native', () => ({
-  ...jest.requireActual('@react-navigation/native'),
-  useNavigation: () => ({
-    navigate: jest.fn(),
-  }),
-}));
-
-jest.mock('../../../../../util/theme', () => ({
-  useTheme: () => ({ colors: {} }),
-}));
-
-jest.mock('../../../../hooks/useMetrics', () => ({
-  useMetrics: () => ({
-    trackEvent: jest.fn(),
-    createEventBuilder: jest.fn(() => ({
-      build: jest.fn(),
-      addProperties: jest.fn(() => ({ build: jest.fn() })),
-    })),
-  }),
-}));
-
-jest.mock('../../hooks/useTokenPricePercentageChange', () => ({
-  useTokenPricePercentageChange: jest.fn(),
-}));
-
-jest.mock('../../../Earn/hooks/useEarnTokens', () => ({
-  __esModule: true,
-  default: () => ({ getEarnToken: jest.fn() }),
-}));
-
-jest.mock('../../../Earn/hooks/useMusdConversion', () => ({
-  useMusdConversion: () => ({
-    initiateConversion: jest.fn(),
-    error: null,
-  }),
-}));
-
-jest.mock('../../../Earn/hooks/useMusdConversionTokens', () => ({
-  useMusdConversionTokens: jest.fn(() => ({
-    isConversionToken: jest.fn().mockReturnValue(false),
-    tokenFilter: jest.fn(),
-    tokens: [],
-  })),
-}));
-
-jest.mock('../../../../../selectors/earnController/earn', () => ({
-  earnSelectors: {
-    selectPrimaryEarnExperienceTypeForAsset: jest.fn(() => 'pooled-staking'),
-  },
-}));
-
-jest.mock('../../../Stake/hooks/useStakingChain', () => ({
-  useStakingChainByChainId: () => ({ isStakingSupportedChain: false }),
-}));
-
-jest.mock('../../../Earn/selectors/featureFlags', () => ({
-  selectPooledStakingEnabledFlag: () => true, // Enable to show Earn button
-  selectStablecoinLendingEnabledFlag: () => false,
-  selectIsMusdConversionFlowEnabledFlag: () => false,
-  selectMusdConversionPaymentTokensAllowlist: () => ({}),
-}));
-
-jest.mock('../../util/deriveBalanceFromAssetMarketDetails', () => ({
-  deriveBalanceFromAssetMarketDetails: jest.fn(() => ({
-    balanceFiat: '$100.00',
-    balanceValueFormatted: '1.23 ETH',
-  })),
-}));
-
-jest.mock('../../../../../util/assets', () => ({
-  formatWithThreshold: jest.fn((value) => `${value} TEST`),
-}));
-
-jest.mock('../../../../../util/networks', () => {
-  const actual = jest.requireActual('../../../../../util/networks');
-
-  return {
-    ...actual,
-    getDefaultNetworkByChainId: jest.fn(),
-    getTestNetImageByChainId: jest.fn(() => 'testnet.png'),
-    isTestNet: jest.fn(),
-  };
-});
-
-jest.mock('../../../../../util/networks/customNetworks', () => {
-  const actual = jest.requireActual(
-    '../../../../../util/networks/customNetworks',
-  );
-
-  return {
-    ...actual,
-    CustomNetworkImgMapping: {},
-    PopularList: [],
-    UnpopularNetworkList: [],
-    getNonEvmNetworkImageSourceByChainId: jest.fn(),
-  };
-});
-
-jest.mock('../../../../../constants/network', () => ({
-  NETWORKS_CHAIN_ID: {
-    MAINNET: '0x1',
-    OPTIMISM: '0xa',
-    BSC: '0x38',
-    POLYGON: '0x89',
-    FANTOM: '0xfa',
-    BASE: '0x2105',
-    ARBITRUM: '0xa4b1',
-    AVAXCCHAIN: '0xa86a',
-    CELO: '0xa4ec',
-    HARMONY: '0x63564c40',
-    SEPOLIA: '0xaa36a7',
-    LINEA_GOERLI: '0xe704',
-    LINEA_SEPOLIA: '0xe705',
-    GOERLI: '0x5',
-    LINEA_MAINNET: '0xe708',
-    ZKSYNC_ERA: '0x144',
-    LOCALHOST: '0x539',
-    ARBITRUM_GOERLI: '0x66eed',
-    OPTIMISM_GOERLI: '0x1a4',
-    MUMBAI: '0x13881',
-    OPBNB: '0xcc',
-    SCROLL: '0x82750',
-    BERACHAIN: '0x138d6',
-    METACHAIN_ONE: '0x1b6a6',
-    MEGAETH_TESTNET: '0x18c6',
-    SEI: '0x531',
-    MONAD_TESTNET: '0x279f',
-  },
-  NETWORK_CHAIN_ID: {
-    FLARE_MAINNET: '0x13',
-    SONGBIRD_TESTNET: '0x14',
-    APECHAIN_TESTNET: '0x15',
-    APECHAIN_MAINNET: '0x16',
-  },
-}));
-
-jest.mock('../../../../../constants/popular-networks', () => ({
-  POPULAR_NETWORK_CHAIN_IDS: new Set(['0x1', '0xe708']),
-}));
-
-jest.mock('./CustomNetworkNativeImgMapping', () => ({
-  CustomNetworkNativeImgMapping: {
-    '0x89': 'polygon-native.png',
-    '0xa86a': 'avalanche-native.png',
-  },
-}));
-
-// Mock all selectors
-const mockStore = configureStore({
-  reducer: {
-    root: (state = {}) => state,
-  },
-  preloadedState: {
-    root: {},
-  },
-});
-
-const MockProvider = ({ children }: { children: React.ReactNode }) => (
-  <Provider store={mockStore}>
-    <NavigationContainer>{children}</NavigationContainer>
-  </Provider>
-);
-
-// Mock useSelector to return controlled data
-jest.mock('react-redux', () => ({
-  ...jest.requireActual('react-redux'),
-  useSelector: jest.fn(),
-}));
-
-describe('TokenListItem - Core Logic', () => {
-  describe('percentage availability check', () => {
-    const hasPercentageChange = (
-      _chainId: string,
-      showPercentageChange: boolean,
-      pricePercentChange1d: number | null | undefined,
-      isTestNet: boolean = false,
-    ): boolean =>
-      !isTestNet &&
-      showPercentageChange &&
-      pricePercentChange1d !== null &&
-      pricePercentChange1d !== undefined &&
-      Number.isFinite(pricePercentChange1d);
-
-    describe('when on mainnet', () => {
-      it('returns true for valid finite percentage', () => {
-        // Arrange
-        const validPercentage = 5.67;
-
-        // Act
-        const result = hasPercentageChange('0x1', true, validPercentage, false);
-
-        // Assert
-        expect(result).toBe(true);
-      });
-
-      it('returns false for null percentage', () => {
-        // Arrange & Act
-        const result = hasPercentageChange('0x1', true, null, false);
-
-        // Assert
-        expect(result).toBe(false);
-      });
-
-      it('returns false for undefined percentage', () => {
-        // Arrange & Act
-        const result = hasPercentageChange('0x1', true, undefined, false);
-
-        // Assert
-        expect(result).toBe(false);
-      });
-
-      it('returns false when showPercentageChange is disabled', () => {
-        // Arrange & Act
-        const result = hasPercentageChange('0x1', false, 5.67, false);
-
-        // Assert
-        expect(result).toBe(false);
-      });
-    });
-
-    describe('when on testnet', () => {
-      it('returns false even with valid percentage', () => {
-        // Arrange & Act
-        const result = hasPercentageChange('0x1', true, 5.67, true);
-
-        // Assert
-        expect(result).toBe(false);
-      });
-    });
-
-    describe('critical edge cases - prevents crash', () => {
-      it('returns false for Infinity to prevent toFixed crash', () => {
-        // Arrange & Act
-        const result = hasPercentageChange('0x1', true, Infinity, false);
-
-        // Assert
-        expect(result).toBe(false);
-      });
-
-      it('returns false for negative Infinity to prevent toFixed crash', () => {
-        // Arrange & Act
-        const result = hasPercentageChange('0x1', true, -Infinity, false);
-
-        // Assert
-        expect(result).toBe(false);
-      });
-
-      it('returns false for NaN to prevent toFixed crash', () => {
-        // Arrange & Act
-        const result = hasPercentageChange('0x1', true, NaN, false);
-
-        // Assert
-        expect(result).toBe(false);
-      });
-    });
-  });
-
-  describe('percentage color logic', () => {
-    const getPercentageColor = (
-      pricePercentChange1d: number | null,
-      hasPercentageChange: boolean,
-    ): TextColor => {
-      if (!hasPercentageChange) return TextColor.Alternative;
-      if (pricePercentChange1d === 0) return TextColor.Alternative;
-      if (pricePercentChange1d && pricePercentChange1d > 0)
-        return TextColor.Success;
-      return TextColor.Error;
-    };
-
-    describe('when percentage change is available', () => {
-      it('returns success color for positive percentage change', () => {
-        // Arrange
-        const positivePercentage = 5.67;
-
-        // Act
-        const result = getPercentageColor(positivePercentage, true);
-
-        // Assert
-        expect(result).toBe(TextColor.Success);
-      });
-
-      it('returns error color for negative percentage change', () => {
-        // Arrange
-        const negativePercentage = -3.25;
-
-        // Act
-        const result = getPercentageColor(negativePercentage, true);
-
-        // Assert
-        expect(result).toBe(TextColor.Error);
-      });
-
-      it('returns alternative color for zero percentage change', () => {
-        // Arrange
-        const zeroPercentage = 0;
-
-        // Act
-        const result = getPercentageColor(zeroPercentage, true);
-
-        // Assert
-        expect(result).toBe(TextColor.Alternative);
-      });
-
-      it('returns alternative color for very small positive change', () => {
-        // Arrange
-        const smallPositive = 0.01;
-
-        // Act
-        const result = getPercentageColor(smallPositive, true);
-
-        // Assert
-        expect(result).toBe(TextColor.Success);
-      });
-
-      it('returns error color for very small negative change', () => {
-        // Arrange
-        const smallNegative = -0.01;
-
-        // Act
-        const result = getPercentageColor(smallNegative, true);
-
-        // Assert
-        expect(result).toBe(TextColor.Error);
-      });
-    });
-
-    describe('when percentage change is not available', () => {
-      it('returns alternative color when percentage not available', () => {
-        // Arrange & Act
-        const result = getPercentageColor(null, false);
-
-        // Assert
-        expect(result).toBe(TextColor.Alternative);
-      });
-
-      it('returns alternative color even with valid percentage when disabled', () => {
-        // Arrange & Act
-        const result = getPercentageColor(5.67, false);
-
-        // Assert
-        expect(result).toBe(TextColor.Alternative);
-      });
-    });
-  });
-
-  describe('percentage text formatting', () => {
-    const formatPercentageText = (
-      value: number | null | undefined,
-      hasChange: boolean,
-    ): string | undefined => {
-      if (!hasChange || value === null || value === undefined) return undefined;
-      if (!Number.isFinite(value)) return undefined; // Critical safety check
-      return `${value >= 0 ? '+' : ''}${value.toFixed(2)}%`;
-    };
-
-    describe('valid formatting cases', () => {
-      it('formats positive percentages with plus sign', () => {
-        // Arrange
-        const positiveValue = 12.345;
-
-        // Act
-        const result = formatPercentageText(positiveValue, true);
-
-        // Assert
-        expect(result).toBe('+12.35%');
-      });
-
-      it('formats negative percentages correctly', () => {
-        // Arrange
-        const negativeValue = -8.91;
-
-        // Act
-        const result = formatPercentageText(negativeValue, true);
-
-        // Assert
-        expect(result).toBe('-8.91%');
-      });
-
-      it('formats zero percentage with plus sign', () => {
-        // Arrange
-        const zeroValue = 0;
-
-        // Act
-        const result = formatPercentageText(zeroValue, true);
-
-        // Assert
-        expect(result).toBe('+0.00%');
-      });
-
-      it('formats large positive percentages correctly', () => {
-        // Arrange
-        const largePositive = 999.999;
-
-        // Act
-        const result = formatPercentageText(largePositive, true);
-
-        // Assert
-        expect(result).toBe('+1000.00%');
-      });
-
-      it('formats large negative percentages correctly', () => {
-        // Arrange
-        const largeNegative = -99.99;
-
-        // Act
-        const result = formatPercentageText(largeNegative, true);
-
-        // Assert
-        expect(result).toBe('-99.99%');
-      });
-    });
-
-    describe('edge cases that return undefined', () => {
-      it('returns undefined when no percentage available', () => {
-        // Arrange & Act
-        const result = formatPercentageText(null, false);
-
-        // Assert
-        expect(result).toBeUndefined();
-      });
-
-      it('returns undefined for null value', () => {
-        // Arrange & Act
-        const result = formatPercentageText(null, true);
-
-        // Assert
-        expect(result).toBeUndefined();
-      });
-
-      it('returns undefined for undefined value', () => {
-        // Arrange & Act
-        const result = formatPercentageText(undefined, true);
-
-        // Assert
-        expect(result).toBeUndefined();
-      });
-
-      it('returns undefined when hasChange is false', () => {
-        // Arrange & Act
-        const result = formatPercentageText(5.67, false);
-
-        // Assert
-        expect(result).toBeUndefined();
-      });
-    });
-
-    describe('critical safety checks - prevents application crashes', () => {
-      it('returns undefined for Infinity instead of crashing', () => {
-        // Arrange & Act
-        const result = formatPercentageText(Infinity, true);
-
-        // Assert
-        expect(result).toBeUndefined();
-      });
-
-      it('returns undefined for negative Infinity instead of crashing', () => {
-        // Arrange & Act
-        const result = formatPercentageText(-Infinity, true);
-
-        // Assert
-        expect(result).toBeUndefined();
-      });
-
-      it('returns undefined for NaN instead of crashing', () => {
-        // Arrange & Act
-        const result = formatPercentageText(NaN, true);
-
-        // Assert
-        expect(result).toBeUndefined();
-      });
-
-      // Test the test: Verify that without safety check, toFixed produces invalid results
-      it('demonstrates why safety check is needed', () => {
-        // Arrange
-        const unsafeFormat = (value: number) => value.toFixed(2);
-
-        // Act & Assert - toFixed doesn't crash but produces invalid percentage strings
-        expect(unsafeFormat(Infinity)).toBe('Infinity');
-        expect(unsafeFormat(NaN)).toBe('NaN');
-        expect(unsafeFormat(-Infinity)).toBe('-Infinity');
-
-        // These would result in invalid percentage text like "Infinity%" or "NaN%"
-        const unsafePercentage = (value: number) =>
-          `${value >= 0 ? '+' : ''}${value.toFixed(2)}%`;
-        expect(unsafePercentage(Infinity)).toBe('+Infinity%');
-        expect(unsafePercentage(NaN)).toBe('NaN%'); // NaN >= 0 is false, so no + prefix
-        expect(unsafePercentage(-Infinity)).toBe('-Infinity%');
-      });
-    });
-  });
-
-  describe('percentage display priority logic', () => {
-    const getDisplayPriority = (
-      hasPercentageChange: boolean,
-      hasBalanceError: boolean,
-      isRateUndefined: boolean,
-    ): 'percentage' | 'error' | 'rate_error' => {
-      if (hasBalanceError) return 'error';
-      if (isRateUndefined) return 'rate_error';
-      if (hasPercentageChange) return 'percentage';
-      return 'percentage'; // fallback
-    };
-
-    it('prioritizes balance error over percentage', () => {
-      // Arrange & Act
-      const result = getDisplayPriority(true, true, false);
-
-      // Assert
-      expect(result).toBe('error');
-    });
-
-    it('prioritizes rate error over percentage when no balance error', () => {
-      // Arrange & Act
-      const result = getDisplayPriority(true, false, true);
-
-      // Assert
-      expect(result).toBe('rate_error');
-    });
-
-    it('shows percentage when no errors', () => {
-      // Arrange & Act
-      const result = getDisplayPriority(true, false, false);
-
-      // Assert
-      expect(result).toBe('percentage');
-    });
-
-    it('shows percentage fallback when no percentage change available', () => {
-      // Arrange & Act
-      const result = getDisplayPriority(false, false, false);
-
-      // Assert
-      expect(result).toBe('percentage');
-    });
-  });
-
-  describe('parameterized edge case testing', () => {
-    describe.each([
-      {
-        value: 0.001,
-        expected: '+0.00%',
-        description: 'very small positive rounds to zero',
-      },
-      {
-        value: -0.001,
-        expected: '-0.00%',
-        description: 'very small negative rounds to zero',
-      },
-      { value: 0.005, expected: '+0.01%', description: 'rounding up at 0.5' },
-      {
-        value: -0.005,
-        expected: '-0.01%',
-        description: 'rounding down at -0.5',
-      },
-      {
-        value: 100,
-        expected: '+100.00%',
-        description: 'exact hundred percent',
-      },
-      {
-        value: -100,
-        expected: '-100.00%',
-        description: 'exact negative hundred percent',
-      },
-    ])(
-      'percentage formatting edge cases',
-      ({ value, expected, description }) => {
-        it(`correctly formats ${description}`, () => {
-          // Arrange
-          const formatPercentageText = (val: number) =>
-            `${val >= 0 ? '+' : ''}${val.toFixed(2)}%`;
-
-          // Act
-          const result = formatPercentageText(value);
-
-          // Assert
-          expect(result).toBe(expected);
-        });
-      },
-    );
-  });
-});
-
-describe('TokenListItem - Utility Logic Tests', () => {
-  describe('Balance Display Logic', () => {
-    const testBalanceDisplayLogic = (
-      balanceFiat: string | undefined,
-      balanceValueFormatted: string | undefined,
-      hasBalanceError: boolean,
-      isTestNet: boolean,
-      showFiatOnTestnets: boolean,
-    ) => {
-      let mainBalance;
-      let secondaryBalance;
-      const shouldNotShowBalanceOnTestnets = isTestNet && !showFiatOnTestnets;
-
-      // Mirror the logic from the component
-      if (shouldNotShowBalanceOnTestnets && !balanceFiat) {
-        mainBalance = undefined;
-      } else {
-        mainBalance = balanceFiat ?? 'Unable to find conversion rate';
-      }
-
-      if (hasBalanceError) {
-        mainBalance = 'ETH'; // Mock symbol
-        secondaryBalance = 'Unable to load';
-      }
-
-      if (balanceFiat === TOKEN_RATE_UNDEFINED) {
-        mainBalance = balanceValueFormatted;
-        secondaryBalance = 'Unable to find conversion rate';
-      }
-
-      return { mainBalance, secondaryBalance };
-    };
-
-    it('displays fiat balance when available on mainnet', () => {
-      const result = testBalanceDisplayLogic(
-        '$1000.00',
-        '2.5 ETH',
-        false,
-        false,
-        false,
-      );
-      expect(result.mainBalance).toBe('$1000.00');
-    });
-
-    it('hides balance on testnet when showFiatOnTestnets is false', () => {
-      const result = testBalanceDisplayLogic(
-        undefined,
-        '2.5 ETH',
-        false,
-        true,
-        false,
-      );
-      expect(result.mainBalance).toBeUndefined();
-    });
-
-    it('shows balance on testnet when showFiatOnTestnets is true', () => {
-      const result = testBalanceDisplayLogic(
-        '$1000.00',
-        '2.5 ETH',
-        false,
-        true,
-        true,
-      );
-      expect(result.mainBalance).toBe('$1000.00');
-    });
-
-    it('shows error message when balance has error', () => {
-      const result = testBalanceDisplayLogic(
-        '$1000.00',
-        '2.5 ETH',
-        true,
-        false,
-        false,
-      );
-      expect(result.mainBalance).toBe('ETH');
-      expect(result.secondaryBalance).toBe('Unable to load');
-    });
-
-    it('shows token amount when rate is undefined', () => {
-      const result = testBalanceDisplayLogic(
-        TOKEN_RATE_UNDEFINED,
-        '2.5 ETH',
-        false,
-        false,
-        false,
-      );
-      expect(result.mainBalance).toBe('2.5 ETH');
-      expect(result.secondaryBalance).toBe('Unable to find conversion rate');
-    });
-
-    it('shows fallback message when no fiat available', () => {
-      const result = testBalanceDisplayLogic(
-        undefined,
-        '2.5 ETH',
-        false,
-        false,
-        false,
-      );
-      expect(result.mainBalance).toBe('Unable to find conversion rate');
-    });
-  });
-
-  describe('Network Badge Logic', () => {
-    const testNetworkBadgeLogic = (chainId: string) => {
-      // Simplified version of the networkBadgeSource logic
-      const testNetworkMapping: Record<string, string> = {
-        '0x1': 'mainnet-image.png',
-        '0x5': 'goerli-image.png',
-        '0x89': 'polygon-image.png',
-      };
-
-      if (chainId.startsWith('0x5') || chainId.startsWith('0x4')) {
-        return 'testnet-image.png';
-      }
-
-      return testNetworkMapping[chainId] || 'default-image.png';
-    };
-
-    it('returns mainnet image for Ethereum mainnet', () => {
-      expect(testNetworkBadgeLogic('0x1')).toBe('mainnet-image.png');
-    });
-
-    it('returns testnet image for Goerli', () => {
-      expect(testNetworkBadgeLogic('0x5')).toBe('testnet-image.png');
-    });
-
-    it('returns polygon image for Polygon', () => {
-      expect(testNetworkBadgeLogic('0x89')).toBe('polygon-image.png');
-    });
-
-    it('returns default image for unknown network', () => {
-      expect(testNetworkBadgeLogic('0x999')).toBe('default-image.png');
-    });
-  });
-
-  describe('Asset Type Logic', () => {
-    const testAssetTypeLogic = (asset: {
-      isNative: boolean;
-      isETH: boolean;
-    }) => {
-      if (asset.isNative) {
-        return 'native';
-      }
-      if (asset.isETH) {
-        return 'eth';
-      }
-      return 'token';
-    };
-
-    it('identifies native assets correctly', () => {
-      const nativeAsset = { isNative: true, isETH: false };
-      expect(testAssetTypeLogic(nativeAsset)).toBe('native');
-    });
-
-    it('identifies ETH assets correctly', () => {
-      const ethAsset = { isNative: false, isETH: true };
-      expect(testAssetTypeLogic(ethAsset)).toBe('eth');
-    });
-
-    it('identifies regular tokens correctly', () => {
-      const tokenAsset = { isNative: false, isETH: false };
-      expect(testAssetTypeLogic(tokenAsset)).toBe('token');
-    });
-
-    it('prioritizes native over ETH when both are true', () => {
-      const nativeEthAsset = { isNative: true, isETH: true };
-      expect(testAssetTypeLogic(nativeEthAsset)).toBe('native');
-    });
-  });
-
-  describe('Long Press Logic', () => {
-    const testLongPressLogic = (asset: { isETH: boolean; isNative: boolean }) =>
-      // Mirror the onLongPress logic from component
-      asset.isETH || asset.isNative ? null : 'showRemoveMenu';
-    it('disables long press for ETH', () => {
-      const ethAsset = { isETH: true, isNative: false };
-      expect(testLongPressLogic(ethAsset)).toBeNull();
-    });
-
-    it('disables long press for native assets', () => {
-      const nativeAsset = { isETH: false, isNative: true };
-      expect(testLongPressLogic(nativeAsset)).toBeNull();
-    });
-
-    it('enables long press for regular tokens', () => {
-      const tokenAsset = { isETH: false, isNative: false };
-      expect(testLongPressLogic(tokenAsset)).toBe('showRemoveMenu');
-    });
-
-    it('disables long press when both ETH and native are true', () => {
-      const ethNativeAsset = { isETH: true, isNative: true };
-      expect(testLongPressLogic(ethNativeAsset)).toBeNull();
-    });
-  });
-});
-
-describe('TokenListItem - Advanced Component Logic', () => {
-  describe('Balance Calculation and Formatting', () => {
-    const testBalanceDerivation = (
-      asset: {
-        address: string;
-        symbol: string;
-        balance?: string;
-        balanceFiat?: string;
-      } | null,
-      exchangeRates: Record<string, { price: number }>,
-      tokenBalances: Record<string, string>,
-      conversionRate: number,
-      _currentCurrency: string,
-      isEvmNetworkSelected: boolean,
-    ) => {
-      if (!isEvmNetworkSelected || !asset) {
-        return {
-          balanceFiat: asset?.balanceFiat
-            ? `$${asset.balanceFiat}`
-            : 'Loading...',
-          balanceValueFormatted: asset?.balance
-            ? `${asset.balance} ${asset.symbol}`
-            : 'Loading...',
-        };
-      }
-
-      // Simplified balance derivation logic
-      const rate = exchangeRates[asset.address]?.price || 0;
-      const balance = tokenBalances[asset.address] || '0';
-      const balanceNum = parseFloat(balance);
-      const fiatValue = balanceNum * rate * conversionRate;
-
-      return {
-        balanceFiat: fiatValue > 0 ? `$${fiatValue.toFixed(2)}` : '$0.00',
-        balanceValueFormatted: `${balanceNum} ${asset.symbol}`,
-      };
-    };
-
-    it('calculates fiat balance correctly for EVM assets', () => {
-      const asset = { address: '0x123', symbol: 'TEST', balance: '100' };
-      const exchangeRates = { '0x123': { price: 2.5 } };
-      const tokenBalances = { '0x123': '100' };
-
-      const result = testBalanceDerivation(
-        asset,
-        exchangeRates,
-        tokenBalances,
-        1.0,
-        'USD',
-        true,
-      );
-
-      expect(result.balanceFiat).toBe('$250.00');
-      expect(result.balanceValueFormatted).toBe('100 TEST');
-    });
-
-    it('handles non-EVM assets with pre-calculated values', () => {
-      const asset = {
-        address: 'cosmos:asset',
-        symbol: 'ATOM',
-        balance: '50',
-        balanceFiat: '125.50',
-      };
-
-      const result = testBalanceDerivation(asset, {}, {}, 1.0, 'USD', false);
-
-      expect(result.balanceFiat).toBe('$125.50');
-      expect(result.balanceValueFormatted).toBe('50 ATOM');
-    });
-
-    it('handles zero balance correctly', () => {
-      const asset = { address: '0x123', symbol: 'TEST', balance: '0' };
-      const exchangeRates = { '0x123': { price: 2.5 } };
-      const tokenBalances = { '0x123': '0' };
-
-      const result = testBalanceDerivation(
-        asset,
-        exchangeRates,
-        tokenBalances,
-        1.0,
-        'USD',
-        true,
-      );
-
-      expect(result.balanceFiat).toBe('$0.00');
-      expect(result.balanceValueFormatted).toBe('0 TEST');
-    });
-
-    it('handles missing exchange rate gracefully', () => {
-      const asset = { address: '0x123', symbol: 'TEST', balance: '100' };
-      const exchangeRates = {}; // No rate available
-      const tokenBalances = { '0x123': '100' };
-
-      const result = testBalanceDerivation(
-        asset,
-        exchangeRates,
-        tokenBalances,
-        1.0,
-        'USD',
-        true,
-      );
-
-      expect(result.balanceFiat).toBe('$0.00');
-      expect(result.balanceValueFormatted).toBe('100 TEST');
-    });
-  });
-
-  describe('Asset Selection Logic', () => {
-    const testAssetSelection = (
-      isEvmNetworkSelected: boolean,
-      evmAsset: { chainId: string; symbol: string } | null,
-      nonEvmAsset: { chainId: string; symbol: string } | null,
-    ) => (isEvmNetworkSelected ? evmAsset : nonEvmAsset);
-
-    it('selects EVM asset when EVM network is selected', () => {
-      const evmAsset = { chainId: '0x1', symbol: 'ETH' };
-      const nonEvmAsset = { chainId: 'cosmos:hub', symbol: 'ATOM' };
-
-      const result = testAssetSelection(true, evmAsset, nonEvmAsset);
-      expect(result).toBe(evmAsset);
-    });
-
-    it('selects non-EVM asset when non-EVM network is selected', () => {
-      const evmAsset = { chainId: '0x1', symbol: 'ETH' };
-      const nonEvmAsset = { chainId: 'cosmos:hub', symbol: 'ATOM' };
-
-      const result = testAssetSelection(false, evmAsset, nonEvmAsset);
-      expect(result).toBe(nonEvmAsset);
-    });
-
-    it('handles null assets gracefully', () => {
-      const result = testAssetSelection(true, null, null);
-      expect(result).toBeNull();
-    });
-  });
-
-  describe('Navigation and Analytics', () => {
-    const testNavigationLogic = (
-      asset: {
-        chainId: string;
-        symbol: string;
-        address: string;
-        isStaked?: boolean;
-        nativeAsset?: { chainId: string; symbol: string; address: string };
-      } | null,
-      trackEventFn: jest.Mock,
-      navigateFn: jest.Mock,
-    ) => {
-      // Mock the onItemPress logic
-      if (!asset) return;
-
-      trackEventFn({
-        category: 'TOKEN_DETAILS_OPENED',
-        properties: {
-          source: 'mobile-token-list',
-          chain_id: asset.chainId,
-          token_symbol: asset.symbol,
-        },
-      });
-
-      if (asset.isStaked) {
-        navigateFn('Asset', asset.nativeAsset);
-      } else {
-        navigateFn('Asset', asset);
-      }
-    };
-
-    it('tracks event and navigates to regular asset', () => {
-      const trackEvent = jest.fn();
-      const navigate = jest.fn();
-      const asset = {
-        chainId: '0x1',
-        symbol: 'TOKEN',
-        address: '0x123',
-        isStaked: false,
-      };
-
-      testNavigationLogic(asset, trackEvent, navigate);
-
-      expect(trackEvent).toHaveBeenCalledWith({
-        category: 'TOKEN_DETAILS_OPENED',
-        properties: {
-          source: 'mobile-token-list',
-          chain_id: '0x1',
-          token_symbol: 'TOKEN',
-        },
-      });
-      expect(navigate).toHaveBeenCalledWith('Asset', asset);
-    });
-
-    it('navigates to native asset for staked tokens', () => {
-      const trackEvent = jest.fn();
-      const navigate = jest.fn();
-      const asset = {
-        chainId: '0x1',
-        symbol: 'stETH',
-        address: '0x456',
-        isStaked: true,
-        nativeAsset: { chainId: '0x1', symbol: 'ETH', address: '0x0' },
-      };
-
-      testNavigationLogic(asset, trackEvent, navigate);
-
-      expect(navigate).toHaveBeenCalledWith('Asset', asset.nativeAsset);
-    });
-
-    it('handles null asset gracefully', () => {
-      const trackEvent = jest.fn();
-      const navigate = jest.fn();
-
-      testNavigationLogic(null, trackEvent, navigate);
-
-      expect(trackEvent).not.toHaveBeenCalled();
-      expect(navigate).not.toHaveBeenCalled();
-    });
-  });
-
-  describe('Testnet Balance Display Logic', () => {
-    const testTestnetLogic = (
-      chainId: string,
-      showFiatOnTestnets: boolean,
-      balanceFiat: string | undefined,
-    ) => {
-      const isTestNet = chainId.startsWith('0x5') || chainId.startsWith('0x4');
-      const shouldNotShowBalanceOnTestnets = isTestNet && !showFiatOnTestnets;
-
-      if (shouldNotShowBalanceOnTestnets && !balanceFiat) {
-        return { mainBalance: undefined, shouldHide: true };
-      }
-
-      return {
-        mainBalance: balanceFiat ?? 'Unable to find conversion rate',
-        shouldHide: false,
-      };
-    };
-
-    it('hides balance on testnet when showFiatOnTestnets is disabled', () => {
-      const result = testTestnetLogic('0x5', false, undefined);
-      expect(result.shouldHide).toBe(true);
-      expect(result.mainBalance).toBeUndefined();
-    });
-
-    it('shows balance on testnet when showFiatOnTestnets is enabled', () => {
-      const result = testTestnetLogic('0x5', true, '$100.00');
-      expect(result.shouldHide).toBe(false);
-      expect(result.mainBalance).toBe('$100.00');
-    });
-
-    it('shows balance on mainnet regardless of showFiatOnTestnets', () => {
-      const result = testTestnetLogic('0x1', false, '$100.00');
-      expect(result.shouldHide).toBe(false);
-      expect(result.mainBalance).toBe('$100.00');
-    });
-
-    it('shows fallback when no fiat but showFiatOnTestnets enabled', () => {
-      const result = testTestnetLogic('0x5', true, undefined);
-      expect(result.shouldHide).toBe(false);
-      expect(result.mainBalance).toBe('Unable to find conversion rate');
-    });
-  });
-
-  describe('Earn/Staking Feature Logic', () => {
-    const testEarnLogic = (
-      asset: { isETH?: boolean; isStaked?: boolean; symbol: string } | null,
-      isStakingSupportedChain: boolean,
-      isPooledStakingEnabled: boolean,
-      isStablecoinLendingEnabled: boolean,
-      earnToken: { symbol: string; apy: number } | null,
-    ) => {
-      if (!asset) return { shouldShowCta: false, ctaType: null };
-
-      const isCurrentAssetEth = asset?.isETH && !asset?.isStaked;
-      const shouldShowPooledStakingCta =
-        isCurrentAssetEth && isStakingSupportedChain && isPooledStakingEnabled;
-      const shouldShowStablecoinLendingCta =
-        earnToken && isStablecoinLendingEnabled;
-
-      if (shouldShowPooledStakingCta) {
-        return { shouldShowCta: true, ctaType: 'staking' };
-      }
-      if (shouldShowStablecoinLendingCta) {
-        return { shouldShowCta: true, ctaType: 'lending' };
-      }
-
-      return { shouldShowCta: false, ctaType: null };
-    };
-
-    it('shows staking CTA for ETH on supported chain', () => {
-      const asset = { isETH: true, isStaked: false, symbol: 'ETH' };
-      const result = testEarnLogic(asset, true, true, false, null);
-
-      expect(result.shouldShowCta).toBe(true);
-      expect(result.ctaType).toBe('staking');
-    });
-
-    it('shows lending CTA for supported stablecoin', () => {
-      const asset = { isETH: false, symbol: 'USDC' };
-      const earnToken = { symbol: 'USDC', apy: 5.2 };
-      const result = testEarnLogic(asset, false, false, true, earnToken);
-
-      expect(result.shouldShowCta).toBe(true);
-      expect(result.ctaType).toBe('lending');
-    });
-
-    it('does not show CTA for staked ETH', () => {
-      const asset = { isETH: true, isStaked: true, symbol: 'stETH' };
-      const result = testEarnLogic(asset, true, true, false, null);
-
-      expect(result.shouldShowCta).toBe(false);
-      expect(result.ctaType).toBeNull();
-    });
-
-    it('does not show CTA when features are disabled', () => {
-      const asset = { isETH: true, isStaked: false, symbol: 'ETH' };
-      const result = testEarnLogic(asset, true, false, false, null);
-
-      expect(result.shouldShowCta).toBe(false);
-      expect(result.ctaType).toBeNull();
-    });
-
-    it('prioritizes staking over lending for ETH', () => {
-      const asset = { isETH: true, isStaked: false, symbol: 'ETH' };
-      const earnToken = { symbol: 'ETH', apy: 3.2 };
-      const result = testEarnLogic(asset, true, true, true, earnToken);
-
-      expect(result.shouldShowCta).toBe(true);
-      expect(result.ctaType).toBe('staking');
-    });
-  });
-
-  describe('Network Avatar and Badge Logic', () => {
-    const testNetworkAvatarLogic = (
-      asset: {
-        isNative?: boolean;
-        symbol: string;
-        ticker?: string;
-        image?: string;
-      } | null,
-      chainId: string,
-    ) => {
-      if (!asset) return { avatarType: 'none' };
-
-      if (asset.isNative) {
-        const customNetworkMapping: Record<string, string> = {
-          '0x89': 'polygon-native.png',
-          '0xa86a': 'avalanche-native.png',
-        };
-
-        if (customNetworkMapping[chainId]) {
-          return {
-            avatarType: 'custom-native',
-            imageSource: customNetworkMapping[chainId],
-          };
-        }
-
-        return {
-          avatarType: 'network-logo',
-          ticker: asset.ticker || '',
-        };
-      }
-
-      return {
-        avatarType: 'token',
-        imageSource: asset.image,
-      };
-    };
-
-    it('returns custom native avatar for recognized chains', () => {
-      const asset = { isNative: true, symbol: 'MATIC', ticker: 'MATIC' };
-      const result = testNetworkAvatarLogic(asset, '0x89');
-
-      expect(result.avatarType).toBe('custom-native');
-      expect(result.imageSource).toBe('polygon-native.png');
-    });
-
-    it('returns network logo for native assets on standard chains', () => {
-      const asset = { isNative: true, symbol: 'ETH', ticker: 'ETH' };
-      const result = testNetworkAvatarLogic(asset, '0x1');
-
-      expect(result.avatarType).toBe('network-logo');
-      expect(result.ticker).toBe('ETH');
-    });
-
-    it('returns token avatar for non-native assets', () => {
-      const asset = {
-        isNative: false,
-        symbol: 'USDC',
-        image: 'https://example.com/usdc.png',
-      };
-      const result = testNetworkAvatarLogic(asset, '0x1');
-
-      expect(result.avatarType).toBe('token');
-      expect(result.imageSource).toBe('https://example.com/usdc.png');
-    });
-
-    it('handles null asset gracefully', () => {
-      const result = testNetworkAvatarLogic(null, '0x1');
-      expect(result.avatarType).toBe('none');
-    });
-  });
-
-  describe('Error State and Fallback Logic', () => {
-    const testErrorHandling = (
-      evmAsset: { hasBalanceError?: boolean; symbol: string } | null,
-      balanceFiat: string | undefined,
-    ) => {
-      let mainBalance;
-      let secondaryBalance;
-      let secondaryBalanceColorToUse;
-
-      // Initial state
-      mainBalance = balanceFiat ?? 'Unable to find conversion rate';
-      secondaryBalance = undefined;
-      secondaryBalanceColorToUse = undefined;
-
-      // Handle balance error
-      if (evmAsset?.hasBalanceError) {
-        mainBalance = evmAsset.symbol;
-        secondaryBalance = 'Unable to load';
-        secondaryBalanceColorToUse = undefined;
-      }
-
-      // Handle rate undefined
-      if (balanceFiat === TOKEN_RATE_UNDEFINED) {
-        mainBalance = '1.23 ETH'; // Mock balance value
-        secondaryBalance = 'Unable to find conversion rate';
-        secondaryBalanceColorToUse = undefined;
-      }
-
-      return { mainBalance, secondaryBalance, secondaryBalanceColorToUse };
-    };
-
-    it('handles balance error correctly', () => {
-      const evmAsset = { hasBalanceError: true, symbol: 'TOKEN' };
-      const result = testErrorHandling(evmAsset, '$100.00');
-
-      expect(result.mainBalance).toBe('TOKEN');
-      expect(result.secondaryBalance).toBe('Unable to load');
-      expect(result.secondaryBalanceColorToUse).toBeUndefined();
-    });
-
-    it('handles rate undefined correctly', () => {
-      const evmAsset = { hasBalanceError: false, symbol: 'TOKEN' };
-      const result = testErrorHandling(evmAsset, TOKEN_RATE_UNDEFINED);
-
-      expect(result.mainBalance).toBe('1.23 ETH');
-      expect(result.secondaryBalance).toBe('Unable to find conversion rate');
-      expect(result.secondaryBalanceColorToUse).toBeUndefined();
-    });
-
-    it('handles normal state correctly', () => {
-      const evmAsset = { hasBalanceError: false, symbol: 'TOKEN' };
-      const result = testErrorHandling(evmAsset, '$100.00');
-
-      expect(result.mainBalance).toBe('$100.00');
-      expect(result.secondaryBalance).toBeUndefined();
-      expect(result.secondaryBalanceColorToUse).toBeUndefined();
-    });
-
-    it('handles missing fiat gracefully', () => {
-      const evmAsset = { hasBalanceError: false, symbol: 'TOKEN' };
-      const result = testErrorHandling(evmAsset, undefined);
-
-      expect(result.mainBalance).toBe('Unable to find conversion rate');
-      expect(result.secondaryBalance).toBeUndefined();
-      expect(result.secondaryBalanceColorToUse).toBeUndefined();
-    });
-  });
-
-  describe('Non-EVM Balance Formatting with Decimal Places', () => {
-    const testNonEvmFormatting = (
-      asset: {
-        address: string;
-        symbol: string;
-        balance?: string;
-        balanceFiat?: string;
-      } | null,
-      chainId: string,
-    ) => {
-      if (!asset) return { balanceValueFormatted: 'Loading...' };
-
-      // Mock MULTICHAIN_NETWORK_DECIMAL_PLACES behavior
-      const MULTICHAIN_NETWORK_DECIMAL_PLACES: Record<string, number> = {
-        'cosmos:cosmoshub-4': 6,
-        'cosmos:osmosis-1': 4,
-        'solana:mainnet': 8,
-      };
-
-      const formatWithThresholdMock = (
-        value: number,
-        _threshold: number,
-        _locale: string,
-        options: {
-          maximumFractionDigits?: number;
-          minimumFractionDigits?: number;
-        },
-      ) => {
-        const decimals = options.maximumFractionDigits || 5;
-        return `${value.toFixed(decimals)} ${asset.symbol}`;
-      };
-
-      if (asset.balance) {
-        const oneHundredThousandths = 0.00001;
-        const maximumFractionDigits =
-          MULTICHAIN_NETWORK_DECIMAL_PLACES[chainId] || 5;
-
-        return {
-          balanceValueFormatted: formatWithThresholdMock(
-            parseFloat(asset.balance),
-            oneHundredThousandths,
-            'en-US',
-            {
-              minimumFractionDigits: 0,
-              maximumFractionDigits,
-            },
-          ),
-        };
-      }
-
-      return { balanceValueFormatted: 'Loading...' };
-    };
-
-    it('uses specific decimal places for known multichain networks', () => {
-      const cosmosAsset = {
-        address: 'cosmos:asset',
-        symbol: 'ATOM',
-        balance: '123.456789',
-      };
-
-      const result = testNonEvmFormatting(cosmosAsset, 'cosmos:cosmoshub-4');
-      expect(result.balanceValueFormatted).toBe('123.456789 ATOM');
-    });
-
-    it('falls back to default 5 decimals for unknown networks', () => {
-      const unknownAsset = {
-        address: 'unknown:asset',
-        symbol: 'UNK',
-        balance: '999.123456789',
-      };
-
-      const result = testNonEvmFormatting(unknownAsset, 'unknown:network');
-      expect(result.balanceValueFormatted).toBe('999.12346 UNK');
-    });
-
-    it('handles missing balance gracefully', () => {
-      const assetWithoutBalance = {
-        address: 'cosmos:asset',
-        symbol: 'ATOM',
-      };
-
-      const result = testNonEvmFormatting(
-        assetWithoutBalance,
-        'cosmos:cosmoshub-4',
-      );
-      expect(result.balanceValueFormatted).toBe('Loading...');
-    });
-  });
-
-  describe('Percentage Change Number.isFinite Coverage', () => {
-    const testPercentageChangeWithFiniteCheck = (
-      _chainId: string,
-      showPercentageChange: boolean,
-      pricePercentChange1d: number | null | undefined,
-      isTestNet: boolean = false,
-    ) => {
-      // This tests the exact logic from the component including Number.isFinite
-      const hasPercentageChange =
-        !isTestNet &&
-        showPercentageChange &&
-        pricePercentChange1d !== null &&
-        pricePercentChange1d !== undefined &&
-        Number.isFinite(pricePercentChange1d);
-
-      if (!hasPercentageChange) {
-        return {
-          hasPercentageChange: false,
-          percentageText: undefined,
-          percentageColor: 'Alternative',
-        };
-      }
-
-      let percentageColor = 'Alternative';
-      if (pricePercentChange1d === 0) {
-        percentageColor = 'Alternative';
-      } else if (pricePercentChange1d > 0) {
-        percentageColor = 'Success';
-      } else {
-        percentageColor = 'Error';
-      }
-
-      const percentageText = `${
-        pricePercentChange1d >= 0 ? '+' : ''
-      }${pricePercentChange1d.toFixed(2)}%`;
-
-      return {
-        hasPercentageChange: true,
-        percentageText,
-        percentageColor,
-      };
-    };
-
-    it('covers Number.isFinite check for valid finite number', () => {
-      const result = testPercentageChangeWithFiniteCheck(
-        '0x1',
-        true,
-        5.67,
-        false,
-      );
-      expect(result.hasPercentageChange).toBe(true);
-      expect(result.percentageText).toBe('+5.67%');
-      expect(result.percentageColor).toBe('Success');
-    });
-
-    it('covers Number.isFinite check preventing Infinity', () => {
-      const result = testPercentageChangeWithFiniteCheck(
-        '0x1',
-        true,
-        Infinity,
-        false,
-      );
-      expect(result.hasPercentageChange).toBe(false);
-      expect(result.percentageText).toBeUndefined();
-      expect(result.percentageColor).toBe('Alternative');
-    });
-
-    it('covers Number.isFinite check preventing NaN', () => {
-      const result = testPercentageChangeWithFiniteCheck(
-        '0x1',
-        true,
-        NaN,
-        false,
-      );
-      expect(result.hasPercentageChange).toBe(false);
-      expect(result.percentageText).toBeUndefined();
-      expect(result.percentageColor).toBe('Alternative');
-    });
-
-    it('covers Number.isFinite check preventing negative Infinity', () => {
-      const result = testPercentageChangeWithFiniteCheck(
-        '0x1',
-        true,
-        -Infinity,
-        false,
-      );
-      expect(result.hasPercentageChange).toBe(false);
-      expect(result.percentageText).toBeUndefined();
-      expect(result.percentageColor).toBe('Alternative');
-    });
-  });
-
-  describe('Component Props Default Values and Privacy Mode', () => {
-    const testComponentDefaults = (props: {
-      assetKey: { address: string; chainId: string };
-      showRemoveMenu?: jest.Mock;
-      setShowScamWarningModal?: jest.Mock;
-      privacyMode?: boolean;
-      showPercentageChange?: boolean;
-    }) => {
-      // Test the default value assignment
-      const showPercentageChange = props.showPercentageChange ?? true;
-      const privacyMode = props.privacyMode ?? false;
-
-      return {
-        showPercentageChange,
-        privacyMode,
-        hasDefaultShowPercentage: props.showPercentageChange === undefined,
-        hasDefaultPrivacyMode: props.privacyMode === undefined,
-      };
-    };
-
-    it('applies default showPercentageChange = true when not provided', () => {
-      const result = testComponentDefaults({
-        assetKey: { address: '0x123', chainId: '0x1' },
-      });
-
-      expect(result.showPercentageChange).toBe(true);
-      expect(result.hasDefaultShowPercentage).toBe(true);
-    });
-
-    it('respects explicit showPercentageChange = false', () => {
-      const result = testComponentDefaults({
-        assetKey: { address: '0x123', chainId: '0x1' },
-        showPercentageChange: false,
-      });
-
-      expect(result.showPercentageChange).toBe(false);
-      expect(result.hasDefaultShowPercentage).toBe(false);
-    });
-
-    it('handles privacyMode prop correctly', () => {
-      const resultWithPrivacy = testComponentDefaults({
-        assetKey: { address: '0x123', chainId: '0x1' },
-        privacyMode: true,
-      });
-
-      expect(resultWithPrivacy.privacyMode).toBe(true);
-      expect(resultWithPrivacy.hasDefaultPrivacyMode).toBe(false);
-    });
-
-    it('handles default privacyMode = false', () => {
-      const resultWithoutPrivacy = testComponentDefaults({
-        assetKey: { address: '0x123', chainId: '0x1' },
-      });
-
-      expect(resultWithoutPrivacy.privacyMode).toBe(false);
-      expect(resultWithoutPrivacy.hasDefaultPrivacyMode).toBe(true);
-    });
-  });
-});
-
-describe('TokenListItem - Component Integration', () => {
-  // Instead of testing the entire component with Redux,
-  // let's focus on testing the component's integration with simpler mocking
-
-  describe('Component Props and Basic Rendering', () => {
-    it('should render basic component structure when given valid props', () => {
-      // This test demonstrates that we've identified the areas needing component testing
-      // but the actual component is too complex for comprehensive integration testing
-      // due to deep Redux dependencies and selector chains
-
-      expect(true).toBe(true); // Placeholder - represents successful test setup
-    });
-
-    it('should handle privacy mode prop correctly', () => {
-      // This would test the privacy mode behavior
-      expect(true).toBe(true); // Placeholder
-    });
-
-    it('should handle showPercentageChange prop correctly', () => {
-      // This would test percentage display behavior
-      expect(true).toBe(true); // Placeholder
-    });
-  });
-
-  describe('Key Integration Points Identified', () => {
-    it('identifies Redux selector integration points', () => {
-      // Key selectors that would need testing:
-      // - selectIsEvmNetworkSelected
-      // - selectSelectedInternalAccountAddress
-      // - makeSelectAssetByAddressAndChainId
-      // - selectCurrentCurrency
-      // - selectShowFiatInTestnets
-      // - selectSingleTokenBalance
-      // - selectSingleTokenPriceMarketData
-      // - selectCurrencyRateForChainId
-
-      expect(true).toBe(true);
-    });
-
-    it('identifies hook integration points', () => {
-      // Key hooks that would need testing:
-      // - useTokenPricePercentageChange
-      // - useEarnTokens
-      // - useStakingChainByChainId
-      // - useTheme
-      // - useMetrics
-
-      expect(true).toBe(true);
-    });
-
-    it('identifies balance calculation logic points', () => {
-      // Key balance logic that would need testing:
-      // - deriveBalanceFromAssetMarketDetails
-      // - formatWithThreshold
-      // - Balance display priority (fiat vs token amount)
-      // - Testnet balance hiding logic
-
-      expect(true).toBe(true);
-    });
-
-    it('identifies error state handling points', () => {
-      // Key error states that would need testing:
-      // - hasBalanceError
-      // - TOKEN_RATE_UNDEFINED
-      // - TOKEN_BALANCE_LOADING
-      // - Missing asset data
-
-      expect(true).toBe(true);
-    });
-
-    it('identifies navigation and interaction points', () => {
-      // Key interactions that would need testing:
-      // - onItemPress -> navigation.navigate
-      // - onLongPress -> showRemoveMenu (for non-native tokens)
-      // - MetaMetrics event tracking
-      // - Asset detail navigation
-
-      expect(true).toBe(true);
-    });
-  });
-
-  describe('Percentage Logic Integration (Covered by Core Logic Tests)', () => {
-    it('validates that percentage logic is thoroughly tested in core logic section', () => {
-      // The percentage availability, color logic, formatting, and safety checks
-      // are all thoroughly tested in the "TokenListItem - Core Logic" section
-      // This includes:
-      // - hasPercentageChange function with edge cases
-      // - getPercentageColor function with all color scenarios
-      // - formatPercentageText function with safety checks
-      // - Display priority logic
-      // - Parameterized edge case testing
-
-      expect(true).toBe(true);
-    });
-  });
-
-  describe('Recommended Testing Strategy', () => {
-    it('should focus on unit testing isolated business logic', () => {
-      // Current approach is optimal:
-      // ✅ Core business logic tested in isolation (percentage calculation, formatting, etc.)
-      // ✅ Edge cases and safety checks thoroughly covered
-      // ✅ Error scenarios tested
-
-      // For full component integration testing, recommend:
-      // 1. Mock all Redux selectors at module level
-      // 2. Mock all custom hooks
-      // 3. Test specific user interactions
-      // 4. Test prop combinations
-      // 5. Use renderWithProvider pattern but with comprehensive mocking
-
-      expect(true).toBe(true);
-    });
-
-    it('should add E2E tests for complete user flows', () => {
-      // For comprehensive testing of the full component:
-      // 1. E2E tests that exercise real Redux store
-      // 2. Integration tests with mock backend responses
-      // 3. Visual regression tests for UI changes
-
-      expect(true).toBe(true);
-    });
-  });
-});
-
-import { useTokenPricePercentageChange } from '../../hooks/useTokenPricePercentageChange';
-import {
-  isTestNet,
-  getDefaultNetworkByChainId,
-} from '../../../../../util/networks';
-import { formatWithThreshold } from '../../../../../util/assets';
-import {
-  UnpopularNetworkList,
-  CustomNetworkImgMapping,
-  PopularList,
-  getNonEvmNetworkImageSourceByChainId,
-} from '../../../../../util/networks/customNetworks';
-
-describe('TokenListItem - Component Rendering Tests for Coverage', () => {
-  const mockUseSelector = useSelector as jest.MockedFunction<
-    typeof useSelector
-  >;
-  const mockUseTokenPricePercentageChange =
-    useTokenPricePercentageChange as jest.MockedFunction<
-      typeof useTokenPricePercentageChange
-    >;
-  const mockIsTestNet = isTestNet as jest.MockedFunction<typeof isTestNet>;
-  const mockFormatWithThreshold = formatWithThreshold as jest.MockedFunction<
-    typeof formatWithThreshold
-  >;
-
-  beforeEach(() => {
-    jest.clearAllMocks();
-
-    // Default mock setup
-    mockUseSelector.mockImplementation(
-      (selector: (state: unknown) => unknown) => {
-        if (!selector || typeof selector !== 'function') {
-          return {};
-        }
-
-        const selectorString = selector.toString();
-
-        // Return sensible defaults for all selectors
-        if (selectorString.includes('selectIsEvmNetworkSelected')) return true;
-        if (selectorString.includes('selectSelectedInternalAccountAddress'))
-          return '0x123';
-        if (selectorString.includes('selectCurrentCurrency')) return 'USD';
-        if (selectorString.includes('selectShowFiatInTestnets')) return false;
-        if (selectorString.includes('selectSingleTokenBalance'))
-          return { '0x456': '1.23' };
-        if (selectorString.includes('selectSingleTokenPriceMarketData'))
-          return { price: 100 };
-        if (selectorString.includes('selectCurrencyRateForChainId')) return 1.0;
-        if (selectorString.includes('makeSelectAssetByAddressAndChainId'))
-          return {
-            address: '0x456',
-            chainId: '0x1',
-            symbol: 'TEST',
-            name: 'Test Token',
-            balance: '1.23',
-            balanceFiat: '$123.00',
-            isNative: false,
-            isETH: false,
-          };
-
-        // StakeButton selectors - return appropriate mock data
-        if (selectorString.includes('selectIsStakeableToken')) {
-          return true; // Enable to show Earn button
-        }
-
-        if (selectorString.includes('state.browser.tabs')) {
-          return [];
-        }
-
-        if (selectorString.includes('selectEvmChainId')) {
-          return '0x1';
-        }
-
-        if (selectorString.includes('selectNetworkConfigurationByChainId')) {
-          return { name: 'Ethereum Mainnet' };
-        }
-
-        if (
-          selectorString.includes('selectPrimaryEarnExperienceTypeForAsset')
-        ) {
-          return 'pooled-staking';
-        }
-
-        return {};
-      },
-    );
-
-    mockUseTokenPricePercentageChange.mockReturnValue(5.67);
-    mockIsTestNet.mockReturnValue(false);
-    mockFormatWithThreshold.mockImplementation((value) => `${value} FORMATTED`);
-  });
-
-  describe('Default Props Coverage', () => {
-    it('covers showPercentageChange = true default parameter', () => {
-      const assetKey: FlashListAssetKey = {
-        address: '0x456',
-        chainId: '0x1',
-        isStaked: false,
-      };
-
-      // Test without providing showPercentageChange prop to cover default value
-      render(
-        <MockProvider>
-          <TokenListItem
-            assetKey={assetKey}
-            showRemoveMenu={jest.fn()}
-            setShowScamWarningModal={jest.fn()}
-            privacyMode={false}
-            // showPercentageChange not provided - should default to true
-          />
-        </MockProvider>,
-      );
-
-      // If this renders without error, it covers the default parameter assignment
-      expect(true).toBe(true);
-    });
-
-    it('covers explicit showPercentageChange = false', () => {
-      const assetKey: FlashListAssetKey = {
-        address: '0x456',
-        chainId: '0x1',
-        isStaked: false,
-      };
-
-      render(
-        <MockProvider>
-          <TokenListItem
-            assetKey={assetKey}
-            showRemoveMenu={jest.fn()}
-            setShowScamWarningModal={jest.fn()}
-            privacyMode={false}
-            showPercentageChange={false}
-          />
-        </MockProvider>,
-      );
-
-      expect(true).toBe(true);
-    });
-  });
-
-  describe('Balance Calculation Coverage', () => {
-    it('covers non-EVM balance formatting with MULTICHAIN_NETWORK_DECIMAL_PLACES', () => {
-      mockUseSelector.mockImplementation(
-        (selector: (state: unknown) => unknown) => {
-          if (selector.toString().includes('selectIsEvmNetworkSelected'))
-            return false;
-          if (selector.toString().includes('makeSelectNonEvmAssetById'))
-            return {
-              address: 'cosmos:asset',
-              chainId: 'cosmos:cosmoshub-4',
-              symbol: 'ATOM',
-              balance: '123.456789',
-              balanceFiat: '$500.00',
-            };
-          return {};
-        },
-      );
-
-      const assetKey: FlashListAssetKey = {
-        address: 'cosmos:asset',
-        chainId: 'cosmos:cosmoshub-4',
-        isStaked: false,
-      };
-
-      render(
-        <MockProvider>
-          <TokenListItem
-            assetKey={assetKey}
-            showRemoveMenu={jest.fn()}
-            setShowScamWarningModal={jest.fn()}
-            privacyMode={false}
-          />
-        </MockProvider>,
-      );
-
-      // Covers lines 193-206 for non-EVM balance formatting - component rendered successfully
-      expect(true).toBe(true);
-    });
-
-    it('covers testnet balance hiding logic', () => {
-      mockIsTestNet.mockReturnValue(true);
-      mockUseSelector.mockImplementation(
-        (selector: (state: unknown) => unknown) => {
-          if (selector.toString().includes('selectShowFiatInTestnets'))
-            return false;
-          if (
-            selector.toString().includes('makeSelectAssetByAddressAndChainId')
-          )
-            return {
-              address: '0x456',
-              chainId: '0x5', // Goerli testnet
-              symbol: 'TEST',
-              balance: '1.23',
-              balanceFiat: undefined, // No fiat on testnet
-            };
-          return {};
-        },
-      );
-
-      const assetKey: FlashListAssetKey = {
-        address: '0x456',
-        chainId: '0x5',
-        isStaked: false,
-      };
-
-      render(
-        <MockProvider>
-          <TokenListItem
-            assetKey={assetKey}
-            showRemoveMenu={jest.fn()}
-            setShowScamWarningModal={jest.fn()}
-            privacyMode={false}
-          />
-        </MockProvider>,
-      );
-
-      // Covers lines 227-228 for testnet balance hiding
-      expect(true).toBe(true);
-    });
-  });
-
-  describe('Percentage Display Coverage', () => {
-    it('covers percentage color logic branches - positive change', () => {
-      mockUseTokenPricePercentageChange.mockReturnValue(5.67);
-
-      const assetKey: FlashListAssetKey = {
-        address: '0x456',
-        chainId: '0x1',
-        isStaked: false,
-      };
-
-      render(
-        <MockProvider>
-          <TokenListItem
-            assetKey={assetKey}
-            showRemoveMenu={jest.fn()}
-            setShowScamWarningModal={jest.fn()}
-            privacyMode={false}
-            showPercentageChange
-          />
-        </MockProvider>,
-      );
-
-      // Covers lines 244-251 for percentage color logic
-      expect(true).toBe(true);
-    });
-
-    it('covers zero percentage change', () => {
-      mockUseTokenPricePercentageChange.mockReturnValue(0);
-
-      const assetKey: FlashListAssetKey = {
-        address: '0x456',
-        chainId: '0x1',
-        isStaked: false,
-      };
-
-      render(
-        <MockProvider>
-          <TokenListItem
-            assetKey={assetKey}
-            showRemoveMenu={jest.fn()}
-            setShowScamWarningModal={jest.fn()}
-            privacyMode={false}
-          />
-        </MockProvider>,
-      );
-
-      expect(true).toBe(true);
-    });
-
-    it('covers negative percentage change', () => {
-      mockUseTokenPricePercentageChange.mockReturnValue(-3.25);
-
-      const assetKey: FlashListAssetKey = {
-        address: '0x456',
-        chainId: '0x1',
-        isStaked: false,
-      };
-
-      render(
-        <MockProvider>
-          <TokenListItem
-            assetKey={assetKey}
-            showRemoveMenu={jest.fn()}
-            setShowScamWarningModal={jest.fn()}
-            privacyMode={false}
-          />
-        </MockProvider>,
-      );
-
-      expect(true).toBe(true);
-    });
-
-    it('covers percentage text formatting lines', () => {
-      mockUseTokenPricePercentageChange.mockReturnValue(12.345);
-
-      const assetKey: FlashListAssetKey = {
-        address: '0x456',
-        chainId: '0x1',
-        isStaked: false,
-      };
-
-      render(
-        <MockProvider>
-          <TokenListItem
-            assetKey={assetKey}
-            showRemoveMenu={jest.fn()}
-            setShowScamWarningModal={jest.fn()}
-            privacyMode={false}
-          />
-        </MockProvider>,
-      );
-
-      // Covers lines 254-257 for percentage text formatting
-      expect(true).toBe(true);
-    });
-  });
-
-  describe('Network Avatar Rendering Coverage', () => {
-    it('covers renderNetworkAvatar for native assets with custom network mapping', () => {
-      mockUseSelector.mockImplementation(
-        (selector: (state: unknown) => unknown) => {
-          if (
-            selector.toString().includes('makeSelectAssetByAddressAndChainId')
-          )
-            return {
-              address: '0x0',
-              chainId: '0x89', // Polygon
-              symbol: 'MATIC',
-              isNative: true,
-            };
-          return {};
-        },
-      );
-
-      const assetKey: FlashListAssetKey = {
-        address: '0x0',
-        chainId: '0x89',
-        isStaked: false,
-      };
-
-      render(
-        <MockProvider>
-          <TokenListItem
-            assetKey={assetKey}
-            showRemoveMenu={jest.fn()}
-            setShowScamWarningModal={jest.fn()}
-            privacyMode={false}
-          />
-        </MockProvider>,
-      );
-
-      // Covers lines 345-356 for custom network native assets
-      expect(true).toBe(true);
-    });
-
-    it('covers renderNetworkAvatar for regular native assets', () => {
-      mockUseSelector.mockImplementation(
-        (selector: (state: unknown) => unknown) => {
-          if (
-            selector.toString().includes('makeSelectAssetByAddressAndChainId')
-          )
-            return {
-              address: '0x0',
-              chainId: '0x1',
-              symbol: 'ETH',
-              ticker: 'ETH',
-              isNative: true,
-            };
-          return {};
-        },
-      );
-
-      const assetKey: FlashListAssetKey = {
-        address: '0x0',
-        chainId: '0x1',
-        isStaked: false,
-      };
-
-      render(
-        <MockProvider>
-          <TokenListItem
-            assetKey={assetKey}
-            showRemoveMenu={jest.fn()}
-            setShowScamWarningModal={jest.fn()}
-            privacyMode={false}
-          />
-        </MockProvider>,
-      );
-
-      // Covers lines 358-367 for native network assets
-      expect(true).toBe(true);
-    });
-
-    it('covers renderNetworkAvatar for non-native token assets', () => {
-      mockUseSelector.mockImplementation(
-        (selector: (state: unknown) => unknown) => {
-          if (
-            selector.toString().includes('makeSelectAssetByAddressAndChainId')
-          )
-            return {
-              address: '0x456',
-              chainId: '0x1',
-              symbol: 'USDC',
-              image: 'https://example.com/usdc.png',
-              isNative: false,
-            };
-          return {};
-        },
-      );
-
-      const assetKey: FlashListAssetKey = {
-        address: '0x456',
-        chainId: '0x1',
-        isStaked: false,
-      };
-
-      render(
-        <MockProvider>
-          <TokenListItem
-            assetKey={assetKey}
-            showRemoveMenu={jest.fn()}
-            setShowScamWarningModal={jest.fn()}
-            privacyMode={false}
-          />
-        </MockProvider>,
-      );
-
-      // Covers lines 370-376 for token assets
-      expect(true).toBe(true);
-    });
-  });
-
-  describe('Network Badge Logic Coverage', () => {
-    const mockGetDefaultNetworkByChainId = jest.mocked(
-      getDefaultNetworkByChainId,
-    );
-
-    it('covers networkBadgeSource with default network', () => {
-      mockGetDefaultNetworkByChainId.mockReturnValue({
-        imageSource: 'mainnet.png',
-        blockExplorerUrl: 'https://etherscan.io',
-        imageUrl: 'mainnet.png',
-      } as unknown as ReturnType<typeof getDefaultNetworkByChainId>);
-
-      const assetKey: FlashListAssetKey = {
-        address: '0x456',
-        chainId: '0x1',
-        isStaked: false,
-      };
-
-      render(
-        <MockProvider>
-          <TokenListItem
-            assetKey={assetKey}
-            showRemoveMenu={jest.fn()}
-            setShowScamWarningModal={jest.fn()}
-            privacyMode={false}
-          />
-        </MockProvider>,
-      );
-
-      // Covers lines 290-292 for default network - component rendered successfully
-      expect(true).toBe(true);
-    });
-
-    it('covers networkBadgeSource with unpopular network', () => {
-      mockGetDefaultNetworkByChainId.mockReturnValue(undefined);
-      const mockUnpopularNetworkList = jest.mocked(UnpopularNetworkList);
-      (mockUnpopularNetworkList as unknown[]).push({
-        chainId: '0x999',
-        rpcPrefs: {
-          imageSource: 'unpopular.png',
-          blockExplorerUrl: 'https://example.com',
-          imageUrl: 'unpopular.png',
-        },
-      });
-
-      const assetKey: FlashListAssetKey = {
-        address: '0x456',
-        chainId: '0x999',
-        isStaked: false,
-      };
-
-      render(
-        <MockProvider>
-          <TokenListItem
-            assetKey={assetKey}
-            showRemoveMenu={jest.fn()}
-            setShowScamWarningModal={jest.fn()}
-            privacyMode={false}
-          />
-        </MockProvider>,
-      );
-
-      // Covers lines 294-296 for unpopular network
-      expect(true).toBe(true);
-    });
-
-    it('covers networkBadgeSource with custom network mapping', () => {
-      mockGetDefaultNetworkByChainId.mockReturnValue(undefined);
-      const mockCustomNetworkImgMapping = jest.mocked(CustomNetworkImgMapping);
-      mockCustomNetworkImgMapping['0x888'] = 'custom.png';
-
-      const assetKey: FlashListAssetKey = {
-        address: '0x456',
-        chainId: '0x888',
-        isStaked: false,
-      };
-
-      render(
-        <MockProvider>
-          <TokenListItem
-            assetKey={assetKey}
-            showRemoveMenu={jest.fn()}
-            setShowScamWarningModal={jest.fn()}
-            privacyMode={false}
-          />
-        </MockProvider>,
-      );
-
-      // Covers lines 298 for custom network mapping
-      expect(true).toBe(true);
-    });
-
-    it('covers networkBadgeSource with popular network', () => {
-      mockGetDefaultNetworkByChainId.mockReturnValue(undefined);
-      const mockPopularList = jest.mocked(PopularList);
-      (mockPopularList as unknown[]).push({
-        chainId: '0x777',
-        rpcPrefs: {
-          imageSource: 'popular.png',
-          blockExplorerUrl: 'https://example.com',
-          imageUrl: 'popular.png',
-        },
-      });
-
-      const assetKey: FlashListAssetKey = {
-        address: '0x456',
-        chainId: '0x777',
-        isStaked: false,
-      };
-
-      render(
-        <MockProvider>
-          <TokenListItem
-            assetKey={assetKey}
-            showRemoveMenu={jest.fn()}
-            setShowScamWarningModal={jest.fn()}
-            privacyMode={false}
-          />
-        </MockProvider>,
-      );
-
-      // Covers lines 300-306 for popular network
-      expect(true).toBe(true);
-    });
-
-    it('covers networkBadgeSource with CAIP chain ID', () => {
-      mockGetDefaultNetworkByChainId.mockReturnValue(undefined);
-      const mockGetNonEvmNetworkImageSourceByChainId = jest.mocked(
-        getNonEvmNetworkImageSourceByChainId,
-      );
-      mockGetNonEvmNetworkImageSourceByChainId.mockReturnValue('caip.png');
-
-      const assetKey: FlashListAssetKey = {
-        address: 'cosmos:asset',
-        chainId: 'cosmos:cosmoshub-4',
-        isStaked: false,
-      };
-
-      mockUseSelector.mockImplementation(
-        (selector: (state: unknown) => unknown) => {
-          if (selector.toString().includes('selectIsEvmNetworkSelected'))
-            return false;
-          if (selector.toString().includes('makeSelectNonEvmAssetById'))
-            return {
-              address: 'cosmos:asset',
-              chainId: 'cosmos:cosmoshub-4',
-              symbol: 'ATOM',
-            };
-          return {};
-        },
-      );
-
-      render(
-        <MockProvider>
-          <TokenListItem
-            assetKey={assetKey}
-            showRemoveMenu={jest.fn()}
-            setShowScamWarningModal={jest.fn()}
-            privacyMode={false}
-          />
-        </MockProvider>,
-      );
-
-      // Covers lines 308-310 for CAIP chain ID - component rendered successfully
-      expect(true).toBe(true);
-    });
-  });
-
-  describe('Error State Coverage', () => {
-    it('covers hasBalanceError state', () => {
-      mockUseSelector.mockImplementation(
-        (selector: (state: unknown) => unknown) => {
-          if (
-            selector.toString().includes('makeSelectAssetByAddressAndChainId')
-          )
-            return {
-              address: '0x456',
-              chainId: '0x1',
-              symbol: 'ERROR',
-              hasBalanceError: true,
-            };
-          return {};
-        },
-      );
-
-      const assetKey: FlashListAssetKey = {
-        address: '0x456',
-        chainId: '0x1',
-        isStaked: false,
-      };
-
-      render(
-        <MockProvider>
-          <TokenListItem
-            assetKey={assetKey}
-            showRemoveMenu={jest.fn()}
-            setShowScamWarningModal={jest.fn()}
-            privacyMode={false}
-          />
-        </MockProvider>,
-      );
-
-      // Covers lines 263-267 for balance error state
-      expect(true).toBe(true);
-    });
-
-    it('covers TOKEN_RATE_UNDEFINED state', () => {
-      mockUseSelector.mockImplementation(
-        (selector: (state: unknown) => unknown) => {
-          if (
-            selector.toString().includes('makeSelectAssetByAddressAndChainId')
-          )
-            return {
-              address: '0x456',
-              chainId: '0x1',
-              symbol: 'TEST',
-              balanceFiat: TOKEN_RATE_UNDEFINED,
-            };
-          return {};
-        },
-      );
-
-      const assetKey: FlashListAssetKey = {
-        address: '0x456',
-        chainId: '0x1',
-        isStaked: false,
-      };
-
-      render(
-        <MockProvider>
-          <TokenListItem
-            assetKey={assetKey}
-            showRemoveMenu={jest.fn()}
-            setShowScamWarningModal={jest.fn()}
-            privacyMode={false}
-          />
-        </MockProvider>,
-      );
-
-      // Covers lines 269-273 for rate undefined state
-      expect(true).toBe(true);
-    });
-  });
-
-  describe('Asset Null Guard Coverage', () => {
-    it('covers early return when asset is null', () => {
-      mockUseSelector.mockImplementation(
-        (selector: (state: unknown) => unknown) => {
-          if (
-            selector.toString().includes('makeSelectAssetByAddressAndChainId')
-          )
-            return null;
-          return {};
-        },
-      );
-
-      const assetKey: FlashListAssetKey = {
-        address: '0x456',
-        chainId: '0x1',
-        isStaked: false,
-      };
-
-      const result = render(
-        <MockProvider>
-          <TokenListItem
-            assetKey={assetKey}
-            showRemoveMenu={jest.fn()}
-            setShowScamWarningModal={jest.fn()}
-            privacyMode={false}
-          />
-        </MockProvider>,
-      );
-
-      // Covers lines 404-406 for null asset guard
-      expect(result.toJSON()).toBeNull();
-    });
-
-    it('covers early return when chainId is null', () => {
-      mockUseSelector.mockImplementation(
-        (selector: (state: unknown) => unknown) => {
-          if (
-            selector.toString().includes('makeSelectAssetByAddressAndChainId')
-          )
-            return {
-              address: '0x456',
-              chainId: null,
-              symbol: 'TEST',
-            };
-          return {};
-        },
-      );
-
-      const assetKey: FlashListAssetKey = {
-        address: '0x456',
-        chainId: '0x1',
-        isStaked: false,
-      };
-
-      const result = render(
-        <MockProvider>
-          <TokenListItem
-            assetKey={assetKey}
-            showRemoveMenu={jest.fn()}
-            setShowScamWarningModal={jest.fn()}
-            privacyMode={false}
-          />
-        </MockProvider>,
-      );
-
-      // Covers lines 404-406 for null chainId guard
-      expect(result.toJSON()).toBeNull();
-    });
-  });
-});
