--- app/components/UI/Tokens/TokenListControlBar/index.ts
@@ -1,1 +1,0 @@
-export { TokenListControlBar } from './TokenListControlBar';
