+++ app/components/UI/shared/BaseControlBar/BaseControlBar.test.tsx
@@ -38,8 +38,7 @@ jest.mock('../../../hooks/useStyles', () => ({
   useStyles: jest.fn(),
 }));
 
-jest.mock('../../Tokens/TokensBottomSheet', () => ({
-  createTokenBottomSheetFilterNavDetails: jest.fn(() => ['TokenFilter', {}]),
+jest.mock('../../Tokens/TokensBottomSheet/TokenSortBottomSheet', () => ({
   createTokensBottomSheetNavDetails: jest.fn(() => ['TokensBottomSheet', {}]),
 }));
