+++ app/components/UI/shared/BaseControlBar/BaseControlBar.tsx
@@ -19,7 +19,7 @@ import { IconName } from '../../../../component-library/components/Icons/Icon';
 import { selectNetworkName } from '../../../../selectors/networkInfos';
 import { selectIsEvmNetworkSelected } from '../../../../selectors/multichainNetworkController';
 import { getNetworkImageSource } from '../../../../util/networks';
-import { createTokensBottomSheetNavDetails } from '../../Tokens/TokensBottomSheet';
+import { createTokensBottomSheetNavDetails } from '../../Tokens/TokenSortBottomSheet/TokenSortBottomSheet';
 import { createNetworkManagerNavDetails } from '../../NetworkManager';
 import { useCurrentNetworkInfo } from '../../../hooks/useCurrentNetworkInfo';
 import {
